better structuring of media storages
