+++ backend/src/media/media_storage/media_storage.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { MediaStorageService } from "./media_storage.service";
+import { MediaStorageService } from "./service/media_storage.service";
 import { MediaStorageController } from "./media_storage.controller";
 import { DbModule } from "../../database/db.module";
