+++ backend/src/media/media_storage/service/storage.interface.ts
@@ -1,0 +1,5 @@
+export interface MediaStorage {
+  save(url: string, buffer: Buffer): Promise<string>;
+  get(url: string): Promise<Buffer>;
+  delete(url: string): Promise<void>;
+}
